Split "quality" plugin into media-specific plugins
