+++ couchpotato/core/media/_base/quality/__init__.py
@@ -1,0 +1,7 @@
+from .main import Quality
+
+
+def autoload():
+    return Quality()
+
+config = []
