+++ couchpotato/core/media/_base/quality/main.py
@@ -1,0 +1,54 @@
+import traceback
+
+from couchpotato import fireEvent, get_db, tryInt, CPLog
+from couchpotato.api import addApiView
+from couchpotato.core.media._base.quality.index import QualityIndex
+from couchpotato.core.plugins.base import Plugin
+
+log = CPLog(__name__)
+
+
+class Quality(Plugin):
+    _database = {
+        'quality': QualityIndex
+    }
+
+    def __init__(self):
+        addApiView('quality.list', self.allView, docs = {
+            'desc': 'List all available qualities',
+            'return': {'type': 'object', 'example': """{
+            'success': True,
+            'list': array, qualities
+}"""}
+        })
+
+        addApiView('quality.size.save', self.saveSize)
+
+    def allView(self, **kwargs):
+
+        return {
+            'success': True,
+            'list': fireEvent('quality.all', merge = True)
+        }
+
+    def saveSize(self, **kwargs):
+
+        try:
+            db = get_db()
+            quality = db.get('quality', kwargs.get('identifier'), with_doc = True)
+
+            if quality:
+                quality['doc'][kwargs.get('value_type')] = tryInt(kwargs.get('value'))
+                db.update(quality['doc'])
+
+            fireEvent('quality.reset_cache')
+
+            return {
+                'success': True
+            }
+        except:
+            log.error('Failed: %s', traceback.format_exc())
+
+        return {
+            'success': False
+        }
