+++ couchpotato/core/media/_base/quality/static/quality.js
