--- couchpotato/core/plugins/quality/__init__.py
@@ -1,5 +1,0 @@
-from .main import QualityPlugin
-
-
-def autoload():
-    return QualityPlugin()
