+++ couchpotato/templates/index.html
@@ -66,7 +66,7 @@
 
 				Quality.setup({
 					'profiles': {{ json_encode(fireEvent('profile.all', single = True)) }},
-					'qualities': {{ json_encode(fireEvent('quality.all', single = True)) }}
+					'qualities': {{ json_encode(fireEvent('quality.all', merge = True)) }}
 				});
 
 				CategoryList.setup({{ json_encode(fireEvent('category.all', single = True)) }});
